+++ lib/Effects/Effects/Throb.h
@@ -2,8 +2,9 @@
 
 void throbEffectLoop(struct EffectDataPacket *effectState, int hue)
 {
+
   const uint16_t loopMs = 3000;
-  const uint8_t loopPosition = effectState->loopPosition % loopMs;
+  const uint16_t loopPosition = effectState->loopPosition % loopMs;
   const float loopPercent = float(loopPosition) / loopMs;
   const uint8_t intensity = abs(cos(PI * .5 * loopPercent * 2) * 255);
